legacy/requirements.txt
Pillow==10.2.0
hachoir~=3.3.0
pillow_heif